+++ tools/gms_tools/overlapping_inundation.py
@@ -13,6 +13,7 @@ from numba import njit
 import geopandas as gpd
 from rasterio.mask import mask
 import sys
+import warnings
 
 class OverlapWindowMerge:
 
@@ -441,7 +442,12 @@ def merge_data(rst_data,
                  data]
 
         del data
-        window_data[row_slice, col_slice] = agg_function(merge)
+        
+        with warnings.catch_warnings():
+            # This `with` block supresses the RuntimeWarning thrown by numpy when aggregating nan values
+            warnings.simplefilter("ignore", category=RuntimeWarning)
+            window_data[row_slice, col_slice] = agg_function(merge)
+        
         window_data[np.isnan(window_data)] = nodata
         del merge